+++ docs/CHANGELOG.md
@@ -15,7 +15,7 @@ Version 1.2.0 (Willow): Not yet released
 - Add `-afs FILENAME` command line argument to specify a settings file when launching
 - Add `Suppress autoswitch` control bind (if held while picking up a weapon, autoswitch is suppressed)
 - Add `SuppressAutoswitchBindAlias` setting to `alpine_settings.ini`
-- Add `$Flag Captures While Stolen` dedicated server config option
+- Add `Flag Captures While Stolen` dedicated server config option
 - Add damage reduction handicap feature
 - Add `mp_handicap` command to configure your damage reduction handicap (0 - 99%)
 - Add `DesiredHandicap` setting to `alpine_settings.ini`
@@ -31,6 +31,7 @@ Version 1.2.0 (Willow): Not yet released
 - Reformatted output of `info` command to contain more useful info
 - Deprecated `sv_connectedclients` command as all info is now in the output of `info`
 - Updated new player joining message printed to ADS server console to include socket, client info, and max RFL version
+- Support weapon stay exemption functionality for any weapons, including stock fusion behaviour and weapons fron TC mods
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)